# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "gpt-terminal-plus API",
    "version": "0.1.0",
    "description": "OpenAPI generated from JSDoc annotations.",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3100",
      "description": "Public base URL"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API_TOKEN"
      }
    },
...
  },
  "paths": {
...
    "/file/read": {
      "post": {
        "operationId": "fileRead",
        "summary": "Read a single file with optional line range",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "filePath": {
                    "type": "string"
                  },
                  "startLine": {
                    "type": "integer",
                    "description": "1-based inclusive start line"
                  },
                  "endLine": {
                    "type": "integer",
                    "description": "1-based inclusive end line"
                  },
                  "encoding": {
                    "type": "string",
                    "default": "utf8"
                  },
                  "maxBytes": {
                    "type": "integer",
                    "default": 1048576
                  }
                },
                "required": [
                  "filePath"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File content",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "filePath": {
                      "type": "string"
                    },
                    "encoding": {
                      "type": "string"
                    },
                    "startLine": {
                      "type": "integer"
                    },
                    "endLine": {
                      "type": "integer"
                    },
                    "truncated": {
                      "type": "boolean"
                    },
                    "content": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "filePath",
                    "content"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        },
        "x-openai-isConsequential": false
      }
    },
...
    "/file/create": {
      "post": {
        "operationId": "fileCreate",
        "summary": "Create or replace a file on the active server",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "filePath": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "backup": {
                    "type": "boolean",
                    "default": true
                  }
                },
                "required": [
                  "filePath"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File created"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "x-openai-isConsequential": false
      }
    },
    "/file/list": {
      "post": {
        "operationId": "fileList",
        "summary": "List files in a directory on the active server",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "directory": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "files": {
                      "type": "object",
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "total": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "offset": {
                          "type": "integer"
                        }
                      }
                    }
                  },
                  "required": [
                    "files"
                  ]
                }
              }
            }
          }
        },
        "x-openai-isConsequential": false
      },
      "get": {
        "operationId": "fileListGet",
        "summary": "List files (GET shim)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "directory",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-openai-isConsequential": false
      }
    },
    "/file/update": {
      "post": {
        "operationId": "fileUpdate",
        "summary": "Regex replace within a file",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "filePath": {
                    "type": "string"
                  },
                  "pattern": {
                    "type": "string"
                  },
                  "replacement": {
                    "type": "string"
                  },
                  "backup": {
                    "type": "boolean",
                    "default": true
                  },
                  "multiline": {
                    "type": "boolean",
                    "default": false
                  }
                },
                "required": [
                  "filePath",
                  "pattern",
                  "replacement"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "x-openai-isConsequential": false
      }
    },
    "/file/amend": {
      "post": {
        "operationId": "fileAmend",
        "summary": "Append content to a file",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "filePath": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "backup": {
                    "type": "boolean",
                    "default": true
                  }
                },
                "required": [
                  "filePath",
                  "content"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Amended"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "x-openai-isConsequential": false
      }
    },
    "/file/diff": {
      "post": {
        "operationId": "fileApplyDiff",
        "summary": "Apply a unified diff using git apply",
        "description": "Validates with `git apply --check`, then applies with `git apply`. Currently supported only for the local server handler.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-openai-isConsequential": false,
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "diff": {
                    "type": "string"
                  },
                  "dryRun": {
                    "type": "boolean",
                    "default": false
                  },
                  "whitespaceNowarn": {
                    "type": "boolean",
                    "default": true
                  }
                },
                "required": [
                  "diff"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Applied"
          },
          "400": {
            "description": "Validation failed"
          }
        }
      }
    },
    "/file/patch": {
      "post": {
        "operationId": "fileApplyPatch",
        "summary": "Apply a structured patch via git apply",
        "description": "Generates a minimal unified diff for the target file and applies it with git apply. Currently supported only for the local server handler.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-openai-isConsequential": false,
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "filePath": {
                    "type": "string"
                  },
                  "search": {
                    "type": "string"
                  },
                  "oldText": {
                    "type": "string"
                  },
                  "replace": {
                    "type": "string"
                  },
                  "all": {
                    "type": "boolean",
                    "default": false
                  },
                  "startLine": {
                    "type": "integer"
                  },
                  "endLine": {
                    "type": "integer"
                  },
                  "dryRun": {
                    "type": "boolean",
                    "default": false
                  },
                  "whitespaceNowarn": {
                    "type": "boolean",
                    "default": true
                  }
                },
                "required": [
                  "filePath",
                  "replace"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Applied"
          },
          "400": {
            "description": "Validation failed"
          }
        }
      }
    },
...
  },
...
}